Update DHT.cpp
A long cable on the sensor has enough capacitance that causes a rounding of the rising edge of each bit.   On the DHT22 the logical bits are where a logic 0 is ~50 uS (DAH) followed by ~25uS (DIT) and a logic 1 is 50uS (DAH) followed by ~70uS (DUH).  The capacitance in the cable length can modify the ratio between DAH and DUH enough that sometimes a DUH is less than a DAH (by one or two counts) which then looks like a DAH DIT or logic 0.    Scaling the DAH count reading by 60% shifts the ratio into a value closer to where DAH==25 counts, DIT==12 counts and DUH==50 counts on a dsPIC compiled with PIC30-GCC.   Now the test if (highCycles > lowCycles)  never fails to detect a logic 1 bit.